requirements.txt
fastapi~=0.104.1
more_itertools
nltk~=3.8.1
Pillow~=10.1.0
python-multipart
torch~=2.1.1
torchvision~=0.16.1
uvicorn[standard]